Move Spawn Test to correct class
